fix: course api (상세, 등록, 전체) 수정

+++ src/main/java/com/umust/dobonglife/domain/course/controller/dto/response/CourseDetailResponse.java
@@ -57,7 +57,7 @@ public record CourseDetailResponse(
     public record BasicInfo(
             String title,
             String subTitle,
-            Double duration,
+            Long duration,
             CourseLevel level,
             List<CourseTheme> themes,
             List<String> tags
