+++ src/main/java/com/umust/dobonglife/domain/course/domain/vo/CourseBasicInfo.java
@@ -25,7 +25,7 @@ public class CourseBasicInfo {
     private String subTitle;
 
     @Column(nullable = false)
-    private Double duration;
+    private Long duration;
 
     @Column(nullable = false)
     @Enumerated(EnumType.STRING)
@@ -33,7 +33,7 @@ public class CourseBasicInfo {
 
     // tag
 
-    public CourseBasicInfo(String title, String subTitle, Double duration, CourseLevel level) {
+    public CourseBasicInfo(String title, String subTitle, Long duration, CourseLevel level) {
         this.title = title;
         this.subTitle = subTitle;
         this.duration = duration;
